Fixed: 스케줄 추가, 삭제시 Confirm 추가

+++ albalog-client/src/components/Modal/DeleteAllScheduleModal.js
@@ -12,22 +12,22 @@ const DeleteAllScheduleModal = ({
   const handleChangeId = (e) => {
     setId(e.target.value);
   };
-  console.log(employeeList);
 
   const handleDeleteScheduleSubmit = async (e) => {
     e.preventDefault();
 
-    try {
-      const response = await client.delete(
-        `/shift/location/${locationId}/employee/${id}/deleteAll`,
-      );
-      handleDeleteModal();
-      if (response.status === 200) {
-        window.location.replace(`/admin/${locationId}/schedule`); // 페이지 이동 후 새로고침
+    const isConfirm = window.confirm('정말 삭제하시겠습니까?');
+
+    if (isConfirm) {
+      try {
+        await client.delete(
+          `/shift/location/${locationId}/employee/${id}/deleteAll`,
+        );
+        handleDeleteModal();
+      } catch (e) {
+        console.error(e);
+        alert('다시 시도해 주세요');
       }
-    } catch (e) {
-      console.error(e);
-      alert('다시 시도해 주세요');
     }
   };
